+++ docs/source/release-notes.rst
@@ -11,6 +11,7 @@ v1.5 Notes
 
 v1.5.9 (2022-01-xx)
 ```````````````````
+- Renamed ``REGCVSG`` to ``REGCV1`` and ``REGCVSG2`` to ``REGCV2``.
 - Multiprocessing now executes on all CPUs that are physical, instead
   of logical. A new package ``psutil`` needs to be installed.
 - Use of ``Selector`` is deprecated.